moving use cases to applicationContext
