+++ web-client/src/applicationContexts/dev.js
@@ -1,6 +1,14 @@
 import awsPersistenceGateway from '../persistence/awsPersistenceGateway';
 
-const API_URL = 'http://localhost:3000/v1';
+import createCase from '../../../business/src/useCases/createCaseProxy';
+import getCaseDetail from '../useCases/getCaseDetail';
+import getCases from '../useCases/getCases';
+import getPetitionsClerkCaseList from '../useCases/getPetitionsClerkCaseList';
+import getUser from '../../../business/src/useCases/getUser';
+import updateCase from '../useCases/updateCase';
+import uploadCasePdfs from '../useCases/uploadCasePdfs';
+
+const API_URL = 'https://efcms-dev.ustc-case-mgmt.flexion.us/v1';
 
 /**
  * Context for the dev environment
@@ -11,6 +19,17 @@ const applicationContext = {
   },
   getPersistenceGateway: () => {
     return awsPersistenceGateway;
+  },
+  getUseCases: () => {
+    return {
+      createCase,
+      getCaseDetail,
+      getCases,
+      getPetitionsClerkCaseList,
+      getUser,
+      updateCase,
+      uploadCasePdfs,
+    };
   },
 };
