+++ web-client/src/applicationContexts/prod.js
@@ -1,5 +1,13 @@
 import awsPersistenceGateway from '../persistence/awsPersistenceGateway';
 
+import createACase from '../../../business/src/useCases/createACaseProxy';
+import getCaseDetail from '../useCases/getCaseDetail';
+import getCases from '../useCases/getCases';
+import getPetitionsClerkCaseList from '../useCases/getPetitionsClerkCaseList';
+import getUser from '../../../business/src/useCases/getUser';
+import updateCase from '../useCases/updateCase';
+import uploadCasePdfs from '../useCases/uploadCasePdfs';
+
 const API_URL = process.env.API_URL || 'http://localhost:8080';
 
 /**
@@ -11,6 +19,15 @@ const applicationContext = {
   },
   getPersistenceGateway: () => {
     return awsPersistenceGateway;
+  },
+  useCases: {
+    createACase,
+    getCaseDetail,
+    getCases,
+    getPetitionsClerkCaseList,
+    getUser,
+    updateCase,
+    uploadCasePdfs,
   },
 };
