+++ web-client/src/index.dev.js
@@ -3,8 +3,8 @@ import Devtools from 'cerebral/devtools';
 
 import app from './app';
 import applicationContext from './applicationContexts/dev';
-import Case from './entities/Case';
-import User from './entities/User';
+// import Case from './entities/Case';
+// import User from './entities/User';
 
 /**
  * Initializes the app with dev environment context
@@ -14,7 +14,7 @@ if (process.env.USTC_DEBUG) {
   debugTools = {
     devtools: Devtools({
       host: 'localhost:8585',
-      allowedTypes: [Blob, User, Case],
+      // allowedTypes: [Blob, User, Case],
     }),
   };
 }
