+++ web-client/src/main/actions/getUser.js
@@ -1,6 +1,7 @@
 import { state } from 'cerebral';
 
-export default async ({ useCases, get, path }) => {
+export default async ({ applicationContext, get, path }) => {
+  const useCases = applicationContext.getUseCases();
   const user = await useCases.getUser(get(state.form.name));
   if (user) return path.success({ user });
   return path.error({
