+++ web-client/src/main/actions/uploadCasePdfs.js
@@ -1,7 +1,8 @@
 import { state } from 'cerebral';
 
 // TODO: rename to upload to case initation PDFs (or something)
-export default async ({ useCases, applicationContext, get, store }) => {
+export default async ({ applicationContext, get, store }) => {
+  const useCases = applicationContext.getUseCases();
   const fileHasUploaded = () => {
     store.set(
       state.petition.uploadsFinished,
