+++ web-client/src/main/index.js
@@ -1,28 +1,10 @@
 import * as sequences from './sequences';
 
-import createACase from '../../../business/src/useCases/createACaseProxy';
-import getCaseDetail from '../useCases/getCaseDetail';
-import getCases from '../useCases/getCases';
-import getPetitionsClerkCaseList from '../useCases/getPetitionsClerkCaseList';
-import getUser from '../../../business/src/useCases/getUser';
-import updateCase from '../useCases/updateCase';
-import uploadCasePdfs from '../useCases/uploadCasePdfs';
-
 /**
  * Main Cerebral module
  */
 export default {
-  providers: {
-    useCases: {
-      createACase,
-      getCaseDetail,
-      getCases,
-      getPetitionsClerkCaseList,
-      getUser,
-      updateCase,
-      uploadCasePdfs,
-    },
-  },
+  providers: {},
   sequences,
   state: {
     currentPage: 'Loading',
@@ -35,11 +17,11 @@ export default {
     petition: {},
     form: {},
     user: {
-      userId: '',
-      // firstName: '',
-      // lastName: '',
-      // token: '',
-      // role: '',
+      userId: 'taxpayer',
+      firstName: 'Tax',
+      lastName: 'Payer',
+      token: 'taxpayer',
+      role: 'taxpayer',
     },
     caseDetail: {},
     cases: [],
